requirements.txt
streamlit
python-dotenv
openai
tweepy
pytrends
APScheduler
pandas
Pillow
faiss-cpu
tqdm